Adds symlinking to the pyarn add command
