+++ src/utils/__tests__/install.test.js
@@ -1,0 +1,8 @@
+// @flow
+import * as install from '../install';
+
+describe('util/install', () => {
+  it('should', () => {
+    expect(true).toEqual(true);
+  });
+});
